Prepare data for harmonization

The function prepares the data for harmonization with Observed Score Equating in a Random Groups design (OSE-RG) by finding all possible opportunities for harmonization in the data.

Usage

ql_prepare(
  data,
  use_relay = F,
  time_relaxation = 0,
  remove_circular_relays = TRUE,
  scale_min_max = list()
)

Arguments

data

A tibble/data.frame containing the data in a format outlined in the details below.

use_relay

A logical indicating whether relay connections should be calculated. Default = FALSE.

time_relaxation

A numeric which specifies how much time is allowed to differ to establish direct connections. Default = 0.

remove_circular_relays

Logical whether circular relays (relays from the same source question to the same target question should be removed). Default is TRUE.

scale_min_max

A named list which contains numeric vectors of length two, e.g. list("Question A" = c(0,5)). The name of the numeric vector should correspond to one of the questions supplied in 'data'.The first element of the numeric vector is the minimum of a scale, the second element is the maximum of a scale. Should be supplied if not all possible response options were used by respondents. See Details for more information.

Value

An object of type questionlink

questionlink_prepare.

Details

OSE-RG can be used when data from two question measuring the same construct are used in the same population at roughly similar times.

The data has to be supplied as a tibble/data.frame in a "long format". The data.frame/tibble has to have the following columns names and variable types:

year: Numeric variable which specifies the year in which the survey was conducted. Can not be of type Date or Factor. If other units of time are needed look here.

question: Character variable which specifies which kind of question was asked.

response: Numeric variable which contains the response values. NAs and missings should be removed. Note that the responses should be recoded so that higher values represent the same meaning on the scale.

(optional) weight: Numeric variable which contains the weight associated with a response.

(optional) population: Character variable which describes the population a response is attributed to. Data should only be harmonized within one population. If no population is supplied all data will be treated to originate from the same population.